requirements.txt
flask
gunicorn
flask-cors
azure
python-dotenv